added api for the ExerciseService, moved all files to their right packages
