+++ backend/src/main/java/com/vule/workoutcalendar/exercise/api/ExerciseServiceApi.java
@@ -1,13 +1,10 @@
 package com.vule.workoutcalendar.exercise.api;
 
 import com.vule.workoutcalendar.exercise.Exercise;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
 
 import java.util.List;
 
-@RequestMapping("/exericses")
-public interface ExerciseControllerApi {
+public interface ExerciseServiceApi {
 
     List<Exercise> findAll();
