+++ backend/src/main/java/com/vule/workoutcalendar/workoutexercise/WorkoutExerciseService.java
@@ -1,6 +1,6 @@
 package com.vule.workoutcalendar.workoutexercise;
 
-import com.vule.workoutcalendar.exercise.ExerciseRepository;
+import com.vule.workoutcalendar.exercise.impl.ExerciseRepository;
 import com.vule.workoutcalendar.workout.WorkoutRepository;
 import com.vule.workoutcalendar.workoutexercise.dto.GroupedExerciseDto;
 import com.vule.workoutcalendar.workoutexercise.dto.WorkoutExerciseDto;
